requirements.txt
Django==5.2.4
asgiref==3.9.1
sqlparse==0.5.3
psycopg2-binary==2.9.10
python-dotenv==1.1.1
factory-boy==3.3.3
Faker==37.4.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10 